renamed ui_pyside to ui for consistency
