+++ python/browser_widget/ui/resources_rc.py
