+++ python/browser_widget/ui/thumbnail_label.py
